pyproject.toml
[project]
name = "mcm-agent"
version = "1.0.0"
description = "Mark – AI-powered marketing campaign generator API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    "google-genai>=1.64.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: live Gemini API tests (run with --integration)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
